backend/requirements.txt
Flask>=2.0
python-dotenv>=0.21
flask-cors>=3.0
waitress>=2.1